test-requirements.txt
pytest~=7.1.3
pytest-cov>=2.8.1
pytest-randomly>=3.12.0
mypy>=1.4.1
types-python-dateutil>=2.8.19
python-dotenv >= 1.0.0
responses >= 0.24.0, < 1.0.0
